+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 * Block: 增加js,css快捷方法
 * isCallback: 回调函数除了闭包,还允许传入字符串,数字
 * Record: reload,saveColl方法使用$this->primaryKey代替id字符串
+* Db: 废弃insertBatch方法，改名为batchInsert
 
 ## 0.9.23 (2017-03-20)